big restructure

+++ rust/src/chapter03/adventure_game/objects/house_master.rs
@@ -1,8 +1,8 @@
 use crate::chapter03::adventure_game::clock::Clock;
-use crate::chapter03::adventure_game::dynamic_type::Obj;
 use crate::chapter03::adventure_game::objects::autonomous_agent;
 use crate::chapter03::adventure_game::objects::autonomous_agent::is_autonomous_agent;
 use crate::chapter03::adventure_game::property_table::Properties;
+use crate::chapter03::dynamic_type::Obj;
 use crate::chapter03::generic_procedures::predicate::declare_superset;
 use crate::chapter03::DebugAny;
 
@@ -13,8 +13,8 @@ pub fn is_house_master(obj: &dyn DebugAny) -> bool {
 pub fn make_house_master(
     name: impl ToString,
     home: Obj,
-    restlessness: f32,
-    irritability: f32,
+    restlessness: f64,
+    irritability: f64,
     clock: &mut Clock,
 ) -> Obj {
     let obj = autonomous_agent::make_autonomous_agent(name, home, restlessness, 1.0 / 10.0, clock);
