+++ rust/src/chapter03/adventure_game/objects/troll.rs
@@ -1,8 +1,8 @@
 use crate::chapter03::adventure_game::clock::Clock;
-use crate::chapter03::adventure_game::dynamic_type::Obj;
 use crate::chapter03::adventure_game::objects::autonomous_agent;
 use crate::chapter03::adventure_game::objects::autonomous_agent::is_autonomous_agent;
 use crate::chapter03::adventure_game::property_table::Properties;
+use crate::chapter03::dynamic_type::Obj;
 use crate::chapter03::generic_procedures::predicate::declare_superset;
 use crate::chapter03::DebugAny;
 
@@ -13,8 +13,8 @@ pub fn is_troll(obj: &dyn DebugAny) -> bool {
 pub fn make_troll(
     name: impl ToString,
     place: Obj,
-    restlessness: f32,
-    hunger: f32,
+    restlessness: f64,
+    hunger: f64,
     clock: &mut Clock,
 ) -> Obj {
     let obj = autonomous_agent::make_autonomous_agent(name, place, restlessness, 1.0 / 10.0, clock);
